# Change trace-level iterators to push predicates down for all queries